Refactor: Improve page structure, SEO, and performance
