+++ app/constants/styles.ts
@@ -1,0 +1,28 @@
+// 柔和的配色方案
+export const colors = {
+  mint: '#c8d6cf',      // 淡薄荷绿色背景
+  sand: '#d7cec0',      // 沙色块
+  white: '#ffffff',     // 白色块
+  darkText: '#3a3a3a',  // 深色文字
+  lightText: '#6a6a6a'  // 浅色文字
+} as const;
+
+// 照片路径配置
+export const photos = {
+  hero: '/family-photos/719BD143-ADD6-4F94-8CFC-BAF43235608A.jpg',
+  dash: '/family-photos/IMG_0875.jpeg',
+  cherry: '/family-photos/L1030065.JPG',
+  jimmy: '/family-photos/jimmy.jpg',
+  tinny: '/family-photos/IMG_3908.jpeg',
+  kelly: '/family-photos/IMG_9664.jpeg',
+  gallery: [
+    '/family-photos/IMG_9200.jpeg',
+    '/family-photos/IMG_9147.jpeg',
+    '/family-photos/IMG_9139.jpeg',
+    '/family-photos/IMG_9071.jpeg',
+    '/family-photos/IMG_1435.jpeg',
+    '/family-photos/IMG_7604.jpeg',
+    '/family-photos/2U2A5498.jpg',
+    '/family-photos/2U2A5506.jpg'
+  ]
+} as const;
