+++ app/constants/texts.ts
@@ -1,0 +1,97 @@
+// 中英文文本内容配置
+export const texts = {
+  homeIntro: {
+    zh: '家，是生活最初的诗篇，是记忆长久的温暖。欢迎来到我们的家庭主页，在这里分享我们的故事、回忆和未来。',
+    en: 'Home is the first poem of life, and the enduring warmth of memories. Welcome to our family homepage, where we share our stories, memories, and future.'
+  },
+  exploreFamily: {
+    zh: '了解我们的家庭 →',
+    en: 'Meet Our Family →'
+  },
+  familyMembers: {
+    zh: '家庭成员',
+    en: 'Family Members'
+  },
+  dash: {
+    name: {
+      zh: 'Dash 黄一孟',
+      en: 'Dash Huang'
+    },
+    title: {
+      zh: '家庭顶梁柱',
+      en: 'Family Pillar'
+    },
+    desc: {
+      zh: '作为家庭的中流砥柱，Dash不仅关心每个家庭成员的需求，也为家庭提供坚实的依靠。他喜欢编程和技术，闲暇时会带领全家一起户外活动。在 <a href="https://x.com/DashHuang" target="_blank" rel="noopener noreferrer" style="color: #1DA1F2; text-decoration: none;">X</a> 和 <a href="https://www.instagram.com/dashhuang/" target="_blank" rel="noopener noreferrer" style="color: #E1306C; text-decoration: none;">Instagram</a> 上分享技术和生活。',
+      en: 'As the backbone of the family, Dash not only cares for the needs of each family member but also provides solid support. He enjoys programming and technology, and in his spare time, leads the family in outdoor activities. Follow him on <a href="https://x.com/DashHuang" target="_blank" rel="noopener noreferrer" style="color: #1DA1F2; text-decoration: none;">X</a> and <a href="https://www.instagram.com/dashhuang/" target="_blank" rel="noopener noreferrer" style="color: #E1306C; text-decoration: none;">Instagram</a> for tech and life.'
+    }
+  },
+  cherry: {
+    name: {
+      zh: 'Cherry 吴智群',
+      en: 'Cherry Wu'
+    },
+    title: {
+      zh: '家庭的灵魂人物',
+      en: 'Soul of the Family'
+    },
+    desc: {
+      zh: 'Cherry是家庭的情感核心，善解人意且充满智慧。她热爱阅读和烹饪，总是能为家人带来美味佳肴和温暖的关怀，是孩子们心中的避风港。',
+      en: 'Cherry is the emotional core of the family, understanding and full of wisdom. She loves reading and cooking, always bringing delicious food and warm care to the family, and is a safe harbor for the children.'
+    }
+  },
+  jimmy: {
+    name: {
+      zh: 'Jimmy',
+      en: 'Jimmy'
+    },
+    title: {
+      zh: '大儿子',
+      en: 'Eldest Son'
+    },
+    desc: {
+      zh: 'Jimmy聪明好学，对科学和自然充满好奇心。他喜欢游戏和网球，总是能带给家人欢乐和惊喜。',
+      en: 'Jimmy is intelligent and eager to learn, with curiosity about science and nature. He enjoys gaming and tennis, always bringing joy and surprises to the family.'
+    }
+  },
+  tinny: {
+    name: {
+      zh: 'Tinny',
+      en: 'Tinny'
+    },
+    title: {
+      zh: '大女儿',
+      en: 'Eldest Daughter'
+    },
+    desc: {
+      zh: 'Tinny充满创造力，热爱艺术和音乐。她喜欢跳舞和钢琴，有着丰富的想象力和细腻的情感。',
+      en: 'Tinny is full of creativity, loving art and music. She enjoys dancing and playing the piano, with rich imagination and delicate emotions.'
+    }
+  },
+  kelly: {
+    name: {
+      zh: 'Kelly',
+      en: 'Kelly'
+    },
+    title: {
+      zh: '二女儿',
+      en: 'Younger Daughter'
+    },
+    desc: {
+      zh: 'Kelly活泼可爱，才3岁，充满好奇心。她最喜欢和姐姐妈妈玩，总是用她天真的笑容感染着全家人。',
+      en: 'Kelly is lively and adorable, only 3 years old, full of curiosity. She loves playing with her sister and mother, always infecting the whole family with her innocent smile.'
+    }
+  },
+  gallery: {
+    zh: '家庭相册',
+    en: 'Family Gallery'
+  },
+  viewGallery: {
+    zh: '查看完整相册 →',
+    en: 'View Full Gallery →'
+  },
+  languageSwitch: {
+    zh: 'English',
+    en: '中文'
+  }
+} as const;
